# Use version of chainbase which implements a slab allocator for small size buffers.