--- a/libraries/chain/controller.cpp
+++ b/libraries/chain/controller.cpp
@@ -2243,7 +2243,7 @@ struct controller_impl {
                section.read_row(rows_for_this_tid, db);
                read_row_count.fetch_add(2u, std::memory_order_relaxed);
 
-               utils_t::preallocate(db, rows_for_this_tid.value);
+               // utils_t::preallocate(db, rows_for_this_tid.value);
                for(size_t idx = 0; idx < rows_for_this_tid.value; idx++) {
                   utils_t::create(db, [this, &section, &more, &t_id](auto& row) {
                      row.t_id = t_id;