--- a/libraries/chaindb/include/chainbase/pinnable_mapped_file.hpp
+++ b/libraries/chaindb/include/chainbase/pinnable_mapped_file.hpp
@@ -5,6 +5,7 @@
 #include <boost/interprocess/sync/file_lock.hpp>
 #include <boost/asio/io_context.hpp>
 #include <boost/container/flat_map.hpp>
+#include <chainbase/small_size_allocator.hpp>
 #include <filesystem>
 #include <vector>
 #include <optional>
@@ -45,7 +46,24 @@ public:
 using segment_manager = bip::managed_mapped_file::segment_manager;
 
 template<typename T>
-using allocator = bip::allocator<T, segment_manager>;
+using segment_allocator_t = bip::allocator<T, segment_manager>;
+
+using byte_segment_allocator_t = segment_allocator_t<char>;
+
+using ss_allocator_t = small_size_allocator<byte_segment_allocator_t>;
+
+// An allocator for objects of type T within the segment_manager
+// -------------------------------------------------------------
+// - If the allocation size (num_objects * sizeof(T)) is less than 512 bytes, it will be routed
+//   through the small size allocator which allocates in batch from the `segment_manager`.
+// - If the allocation size (num_objects * sizeof(T)) is greater than 512 bytes, the allocator
+//   will allocate directly from the segment manager.
+// - the 512 bytes limit is derived from the template parameters of `small_size_allocator`
+//   (size_t num_allocators = 64, size_t size_increment = 8)
+// - emulates the API of `bip::allocator<T, segment_manager>`
+// ---------------------------------------------------------------------------------------------
+template<typename T>
+using allocator = object_allocator<T, ss_allocator_t>; 
 
 class pinnable_mapped_file {
    public:
@@ -66,19 +84,22 @@ class pinnable_mapped_file {
       segment_manager* get_segment_manager() const { return _segment_manager;}
       size_t           check_memory_and_flush_if_needed();
 
+      static ss_allocator_t* get_small_size_allocator(std::byte* seg_mgr);
+
       template<typename T>
       static std::optional<allocator<T>> get_allocator(void *object) {
          if (!_segment_manager_map.empty()) {
             auto it = _segment_manager_map.upper_bound(object);
             if(it == _segment_manager_map.begin())
                return {};
-            auto [seg_start, seg_end] = *(--it);
+            auto& [seg_start, seg_info] = *(--it);
             // important: we need to check whether the pointer is really within the segment, as shared objects'
             // can also be created on the stack (in which case the data is actually allocated on the heap using
             // std::allocator). This happens for example when `shared_cow_string`s are inserted into a bip::multimap,
             // and temporary pairs are created on the stack by the bip::multimap code.
-            if (object < seg_end)
-               return allocator<T>(reinterpret_cast<segment_manager *>(seg_start));
+            if (object < seg_info.seg_end) {
+               return std::optional<allocator<T>>{allocator<T>(get_small_size_allocator(static_cast<std::byte*>(seg_start)))};
+            }
          }
          return {};
       }
@@ -114,12 +135,31 @@ class pinnable_mapped_file {
 
       static std::vector<pinnable_mapped_file*>     _instance_tracker;
 
-      using segment_manager_map_t = boost::container::flat_map<void*, void *>;
+      struct seg_info_t { void* seg_end; };
+      using segment_manager_map_t = boost::container::flat_map<void*, seg_info_t>;
       static segment_manager_map_t                  _segment_manager_map;
 
       constexpr static unsigned                     _db_size_multiple_requirement = 1024*1024; //1MB
       constexpr static size_t                       _db_size_copy_increment       = 1024*1024*1024; //1GB
 };
+
+// There can be at most one `small_size_allocator` per `segment_manager` (hence the `assert` below).
+// There is none created if the pinnable_mapped_file is read-only.
+// ----------------------------------------------------------------------------------------------------
+template <class backing_allocator>
+auto make_small_size_allocator(segment_manager* seg_mgr) {
+   assert(pinnable_mapped_file::get_small_size_allocator((std::byte*)seg_mgr) == nullptr);
+   byte_segment_allocator_t byte_allocator(seg_mgr);
+   return new (seg_mgr->allocate(sizeof(ss_allocator_t))) ss_allocator_t(byte_allocator);
+}
+
+// Create an allocator for a specific object type. 
+// pointer can be to the segment manager, or any object contained within.
+// ---------------------------------------------------------------------
+template <class T>
+auto make_allocator(void* seg_mgr) {
+   return *pinnable_mapped_file::get_allocator<T>(seg_mgr);
+}
 
 std::istream& operator>>(std::istream& in, pinnable_mapped_file::map_mode& runtime);
 std::ostream& operator<<(std::ostream& osm, pinnable_mapped_file::map_mode m);