Resolves #1049.

As the issue describes, the current chainbase does a lot of small buffer allocations directly through the bip allocator which hits the segment manager directly. Not only this is inefficient (the rbtree of free buffers need to be rebalanced for every alloc/free), it also has a higher memory overhead. 

The slab allocators for small buffers batches allocations of small buffers into larger blocks, reducing the fragmentattion within the shared memory segment, and providing better efficiency and a smaller memory footprint.

For comparison, here are some results comparing snapshot loading on my linux box (release build)

#### snapshot from Aug 11, 2024

|                  | time (average) | memory usage | memory saving |
|---------------|----------------------|------------------------|------------------------|
| main branch | 55s | 23.67GB | - |
| this branch | 28s | 22.41GB | 5.3% |

#### snapshot from Sept 29, 2024

|                  | time (average) | memory usage | memory saving |
|---------------|----------------------|------------------------|------------------------|
| main branch | 392s | 65.30GB | - |
| this branch | 331s | 63.12GB | 3.3% |

I also expect a slight performance increase for processing transactions, as the `_undo_stack` deque allocations used to go through the segment allocator as well (for blocks of 16 `undo_state`), and these allocations are now batched.