@ericpassmore 2024-12-16
Note:start
category: System Enhancements
component: Internal
summary: Improve chainbase to use utilize slab allocator for small buffers. Reduces memory contention and frequency of rebalancing. 
Note:end